admin login and user login added successfully

+++ src/main/java/com/ecom/controller/HomeController.java
@@ -39,7 +39,7 @@ public class HomeController {
 
 	@Autowired
 	private UserService userService;
-	
+
 	@Autowired
 	private Productservice productservice;
 
@@ -75,34 +75,36 @@ public class HomeController {
 		m.addAttribute("product", productById);
 		return "view_product";
 	}
-	
-	@PostMapping("/saveUser")
-	public String saveUser(@ModelAttribute  UserDtls user, @RequestParam("img") MultipartFile file,HttpSession session
-			) throws IOException
-	{
-	String imageName=file.isEmpty() ? "default.jpg" : file.getOriginalFilename();
-		user.setProfileImage(imageName);
-		UserDtls saveUser=userService.saveUser(user);
-		
-		if(!ObjectUtils.isEmpty(saveUser))
-		{
-		if(!file.isEmpty())
-		{
-			File saveFile = new ClassPathResource("static/img").getFile();
-			
-			Path path = Paths.get(saveFile.getAbsolutePath()+File.separator + "profile_img" + 
-			File.separator + file.getOriginalFilename());
-//			System.out.println(path);
-			Files.copy(file.getInputStream(), path,StandardCopyOption.REPLACE_EXISTING );
-		}
-		session.setAttribute("succMsg", "Saved successfully");
-		}
-		
-		else
-		{
-			session.setAttribute("errorMsg", "Something wrong on server");
 
+	@PostMapping("/saveUser")
+	public String saveUser(@ModelAttribute UserDtls user, @RequestParam("img") MultipartFile file, HttpSession session)
+			throws IOException {
+
+		Boolean existsEmail = userService.existsEmail(user.getEmail());
+
+		if (existsEmail) {
+			session.setAttribute("errorMsg", "Email already exist");
+		} else {
+			String imageName = file.isEmpty() ? "default.jpg" : file.getOriginalFilename();
+			user.setProfileImage(imageName);
+			UserDtls saveUser = userService.saveUser(user);
+
+			if (!ObjectUtils.isEmpty(saveUser)) {
+				if (!file.isEmpty()) {
+					File saveFile = new ClassPathResource("static/img").getFile();
+
+					Path path = Paths.get(saveFile.getAbsolutePath() + File.separator + "profile_img" + File.separator
+							+ file.getOriginalFilename());
+
+//					System.out.println(path);
+					Files.copy(file.getInputStream(), path, StandardCopyOption.REPLACE_EXISTING);
+				}
+				session.setAttribute("succMsg", "Register successfully");
+			} else {
+				session.setAttribute("errorMsg", "something wrong on server");
+			}
 		}
+
 		return "redirect:/register";
 	}
